+++ contracts/starfleit_pair/src/error.rs
@@ -29,4 +29,13 @@ pub enum ContractError {
 
     #[error("Expired deadline")]
     ExpiredDeadline {},
+
+    #[error("Max slippage assertion")]
+    MaxSlippageAssertion {},
+
+    #[error("More initial liquidity needed ({min_lp_token} > {given_lp})")]
+    MinimumLiquidityAmountError {
+        min_lp_token: String,
+        given_lp: String,
+    },
 }