+++ contracts/starfleit_pair/src/testing.rs
@@ -4,11 +4,12 @@ use crate::contract::{
 };
 use crate::error::ContractError;
 use starfleit::mock_querier::mock_dependencies;
+use std::str::FromStr;
 
 use cosmwasm_std::testing::{mock_env, mock_info, MOCK_CONTRACT_ADDR};
 use cosmwasm_std::{
-    attr, coins, to_binary, BankMsg, Coin, CosmosMsg, Decimal, Reply, ReplyOn, Response, StdError,
-    SubMsg, SubMsgResponse, SubMsgResult, Uint128, WasmMsg,
+    attr, to_binary, BankMsg, Coin, CosmosMsg, Decimal, Reply, ReplyOn, Response, StdError, SubMsg,
+    SubMsgResponse, SubMsgResult, Uint128, WasmMsg,
 };
 use cw20::{Cw20ExecuteMsg, Cw20ReceiveMsg, MinterResponse};
 use starfleit::asset::{Asset, AssetInfo, PairInfo};
@@ -102,7 +103,7 @@ fn proper_initialization() {
 fn provide_liquidity() {
     let mut deps = mock_dependencies(&[Coin {
         denom: "afet".to_string(),
-        amount: Uint128::from(200u128),
+        amount: Uint128::from(1_100u128),
     }]);
 
     deps.querier.with_token_balances(&[
@@ -147,6 +148,48 @@ fn provide_liquidity() {
 
     let _res = reply(deps.as_mut(), mock_env(), reply_msg).unwrap();
 
+    // should raise MinimumLiquidityAmountError with insufficient initial liquidity
+    let msg = ExecuteMsg::ProvideLiquidity {
+        assets: [
+            Asset {
+                info: AssetInfo::Token {
+                    contract_addr: "asset0000".to_string(),
+                },
+                amount: Uint128::from(1u128),
+            },
+            Asset {
+                info: AssetInfo::NativeToken {
+                    denom: "afet".to_string(),
+                },
+                amount: Uint128::from(1u128),
+            },
+        ],
+        receiver: None,
+        deadline: None,
+        slippage_tolerance: None,
+    };
+    let env = mock_env();
+    let info = mock_info(
+        "addr0000",
+        &[Coin {
+            denom: "afet".to_string(),
+            amount: Uint128::from(1u128),
+        }],
+    );
+
+    let res = execute(deps.as_mut(), env, info, msg).unwrap_err();
+
+    match res {
+        ContractError::MinimumLiquidityAmountError {
+            min_lp_token,
+            given_lp,
+        } => {
+            assert_eq!(min_lp_token, "1000");
+            assert_eq!(given_lp, "1");
+        }
+        _ => panic!("Must return MinimumLiquidityAmountError"),
+    }
+
     // successfully provide liquidity for the exist pool
     let msg = ExecuteMsg::ProvideLiquidity {
         assets: [
@@ -154,17 +197,18 @@ fn provide_liquidity() {
                 info: AssetInfo::Token {
                     contract_addr: "asset0000".to_string(),
                 },
-                amount: Uint128::from(100u128),
+                amount: Uint128::from(1_100u128),
             },
             Asset {
                 info: AssetInfo::NativeToken {
                     denom: "afet".to_string(),
                 },
-                amount: Uint128::from(100u128),
+                amount: Uint128::from(1_100u128),
             },
         ],
         receiver: None,
         deadline: None,
+        slippage_tolerance: None,
     };
 
     let env = mock_env();
@@ -172,12 +216,27 @@ fn provide_liquidity() {
         "addr0000",
         &[Coin {
             denom: "afet".to_string(),
-            amount: Uint128::from(100u128),
+            amount: Uint128::from(1_100u128),
         }],
     );
     let res = execute(deps.as_mut(), env, info, msg).unwrap();
-    let transfer_from_msg = res.messages.get(0).expect("no message");
-    let mint_msg = res.messages.get(1).expect("no message");
+    let liquidity_to_contract_msg = res.messages.get(0).expect("no message");
+    let transfer_from_msg = res.messages.get(1).expect("no message");
+    let mint_msg = res.messages.get(2).expect("no message");
+
+    assert_eq!(
+        liquidity_to_contract_msg,
+        &SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
+            contract_addr: "liquidity0000".to_string(),
+            msg: to_binary(&Cw20ExecuteMsg::Mint {
+                recipient: MOCK_CONTRACT_ADDR.to_string(),
+                amount: 1_000u128.into(),
+            })
+            .unwrap(),
+            funds: vec![],
+        }))
+    );
+
     assert_eq!(
         transfer_from_msg,
         &SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
@@ -185,7 +244,7 @@ fn provide_liquidity() {
             msg: to_binary(&Cw20ExecuteMsg::TransferFrom {
                 owner: "addr0000".to_string(),
                 recipient: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Uint128::from(100u128),
+                amount: Uint128::from(1_100u128),
             })
             .unwrap(),
             funds: vec![],
@@ -204,8 +263,8 @@ fn provide_liquidity() {
         }))
     );
 
-    // provide more liquidity 1:2, which is not proportional to 1:1,
-    // then it must accept 1:1 and treat left amount as donation
+    // providing liquidity with a ratio exceeding the specified slippage tolerance
+    // should return MaxSlippageAssertion
     deps.querier.with_balance(&[(
         &MOCK_CONTRACT_ADDR.to_string(),
         vec![Coin {
@@ -244,6 +303,7 @@ fn provide_liquidity() {
         ],
         receiver: Some("staking0000".to_string()), // try changing receiver
         deadline: None,
+        slippage_tolerance: Some(Decimal::from_str("0.005").unwrap()),
     };
 
     let env = mock_env();
@@ -255,19 +315,68 @@ fn provide_liquidity() {
         }],
     );
 
-    // only accept 100, then 50 share will be generated with 100 * (100 / 200)
-    let res: Response = execute(deps.as_mut(), env, info, msg).unwrap();
-    let refund_msg = res.messages.get(0).expect("no message");
-    let transfer_from_msg = res.messages.get(1).expect("no message");
-    let mint_msg = res.messages.get(2).expect("no message");
+    let res = execute(deps.as_mut(), env, info, msg).unwrap_err();
+    match res {
+        ContractError::MaxSlippageAssertion { .. } => (),
+        _ => panic!("MaxSlippageAssertion should be raised"),
+    }
 
-    assert_eq!(
-        refund_msg,
-        &SubMsg::new(CosmosMsg::Bank(BankMsg::Send {
-            to_address: "addr0000".to_string(),
-            amount: coins(100u128, "afet".to_string())
-        }))
+    // providing liquidity at a rate that is not equal to the existing one
+    // should refund the remained amount of one side's assets
+    deps.querier.with_balance(&[(
+        &MOCK_CONTRACT_ADDR.to_string(),
+        vec![Coin {
+            denom: "afet".to_string(),
+            amount: Uint128::from(
+                100u128 + 200u128, /* user deposit must be pre-applied */
+            ),
+        }],
+    )]);
+
+    deps.querier.with_token_balances(&[
+        (
+            &"liquidity0000".to_string(),
+            &[(&MOCK_CONTRACT_ADDR.to_string(), &Uint128::from(100u128))],
+        ),
+        (
+            &"asset0000".to_string(),
+            &[(&MOCK_CONTRACT_ADDR.to_string(), &Uint128::from(200u128))],
+        ),
+    ]);
+
+    let msg = ExecuteMsg::ProvideLiquidity {
+        assets: [
+            Asset {
+                info: AssetInfo::Token {
+                    contract_addr: "asset0000".to_string(),
+                },
+                amount: Uint128::from(100u128),
+            },
+            Asset {
+                info: AssetInfo::NativeToken {
+                    denom: "afet".to_string(),
+                },
+                amount: Uint128::from(100u128),
+            },
+        ],
+        receiver: Some("staking0000".to_string()), // try changing receiver
+        deadline: None,
+        slippage_tolerance: Some(Decimal::from_str("0.05").unwrap()),
+    };
+
+    let env = mock_env();
+    let info = mock_info(
+        "addr0000",
+        &[Coin {
+            denom: "afet".to_string(),
+            amount: Uint128::from(100u128),
+        }],
     );
+
+    let res: Response = execute(deps.as_mut(), env, info, msg).unwrap();
+    let transfer_from_msg = res.messages.get(0).expect("no message");
+    let mint_msg = res.messages.get(1).expect("no message");
+
     assert_eq!(
         transfer_from_msg,
         &SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
@@ -312,6 +421,7 @@ fn provide_liquidity() {
         ],
         receiver: None,
         deadline: None,
+        slippage_tolerance: Some(Decimal::from_str("0.005").unwrap()),
     };
 
     let env = mock_env();
@@ -369,6 +479,7 @@ fn provide_liquidity() {
         ],
         receiver: None,
         deadline: None,
+        slippage_tolerance: Some(Decimal::from_str("0.05").unwrap()),
     };
 
     let env = mock_env();
@@ -575,6 +686,25 @@ fn withdraw_liquidity() {
     let info = mock_info("liquidity0000", &[]);
     let err = execute(deps.as_mut(), env, info, msg).unwrap_err();
     assert_eq!(err, ContractError::ExpiredDeadline {})
+}
+
+#[test]
+fn failed_reply_with_unknown_id() {
+    let mut deps = mock_dependencies(&[]);
+
+    let res = reply(
+        deps.as_mut(),
+        mock_env(),
+        Reply {
+            id: 9,
+            result: SubMsgResult::Ok(SubMsgResponse {
+                events: vec![],
+                data: Some(vec![].into()),
+            }),
+        },
+    );
+
+    assert_eq!(res, Err(StdError::generic_err("invalid reply msg")))
 }
 
 #[test]